+++ frontend/package.json
@@ -48,6 +48,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.0",
+    "react-markdown": "^9.0.1",
     "react-resizable-panels": "^2.1.2",
     "react-virtuoso": "^4.10.4",
     "recharts": "^2.12.7",
