+++ frontend/src/routes/_account/workspaces/$workspaceId/threads.$threadId.tsx
@@ -75,6 +75,7 @@ import {
   PanelRightIcon,
 } from "lucide-react";
 import React from "react";
+import ReactMarkdown from "react-markdown";
 import { useStore } from "zustand";
 
 export const Route = createFileRoute(
@@ -146,38 +147,40 @@ function Message({
           className="mb-3 mt-3 dark:bg-neutral-700"
           orientation="horizontal"
         />
-        <div>{message.body}</div>
+        <div>
+          <ReactMarkdown>{message.body}</ReactMarkdown>
+        </div>
       </div>
     </div>
   );
 }
 
-function SettingThreadLabel() {
-  return (
-    <div className="mr-1 flex">
-      <svg
-        className="h-3 w-3 animate-spin text-indigo-500"
-        fill="none"
-        viewBox="0 0 24 24"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <circle
-          className="opacity-25"
-          cx="12"
-          cy="12"
-          r="10"
-          stroke="currentColor"
-          strokeWidth="4"
-        ></circle>
-        <path
-          className="opacity-75"
-          d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-          fill="currentColor"
-        ></path>
-      </svg>
-    </div>
-  );
-}
+// function SettingThreadLabel() {
+//   return (
+//     <div className="mr-1 flex">
+//       <svg
+//         className="h-3 w-3 animate-spin text-indigo-500"
+//         fill="none"
+//         viewBox="0 0 24 24"
+//         xmlns="http://www.w3.org/2000/svg"
+//       >
+//         <circle
+//           className="opacity-25"
+//           cx="12"
+//           cy="12"
+//           r="10"
+//           stroke="currentColor"
+//           strokeWidth="4"
+//         ></circle>
+//         <path
+//           className="opacity-75"
+//           d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
+//           fill="currentColor"
+//         ></path>
+//       </svg>
+//     </div>
+//   );
+// }
 
 function ThreadLabels({
   threadId,
@@ -320,11 +323,7 @@ function ThreadLabels({
         <DropdownMenu>
           <DropdownMenuTrigger asChild>
             <Button className="h-7 border-dashed" size="sm" variant="outline">
-              {threadLabelMutation.isPending ? (
-                <SettingThreadLabel />
-              ) : (
-                <PlusIcon className="mr-1 h-3 w-3" />
-              )}
+              <PlusIcon className="mr-1 h-3 w-3" />
               Add
             </Button>
           </DropdownMenuTrigger>
